+++ src/index.ts
@@ -10,6 +10,6 @@ const tokenizer = new Tokenizer();
 const requestWrapper = new RequestWrapper();
 const streamFactory = new StreamFactory(cacheFactory, requestWrapper);
 const requestManager = new RequestManager(streamFactory, tokenizer);
-const warden = new Warden(requestManager, requestWrapper);
+const warden = new Warden(requestManager, requestWrapper, cacheFactory);
 
 export default warden;